requirements.txt
pandas>=1.3.0
PyYAML>=5.4
pdfplumber>=0.5.28